Attempt to avoid per-test transactions
This doesn't work, but it gestures in the right direction. If each test
is wrapped in a transaction, you cannot run `migrate` inside the test
successfully (at least under certain circumstances?)

Yet, transactional tests are themselves desirable, and possibly, with
pytest-django, inevitable. Do we need to find a way to apply migrations
in a transaction? Is that even tractable? Or do we need to find a way to
untransact DB tests?